% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diagnose_plot.R
\name{get_diagnose_plot}
\alias{get_diagnose_plot}
\title{CDDR Diagnostic: the plotted comparison of the probability of each causal discovery outcome as a function of sample size less than dataset size N.}
\usage{
get_diagnose_plot(
  cddr,
  outcome_columns,
  dataset_names,
  is_test = T,
  nsubsamp = 100,
  ci_level = 0.95
)
}
\arguments{
\item{cddr}{A dataframe of the estimated causal discovery outcome rates for each subsample size considered, obtained by running run_procedure}

\item{outcome_columns}{A vector the names corresponding to the outcome rates in the cddr dataframe}

\item{dataset_names}{A vector of names corresponding to what one would want to label the possible causal discovery outcome rates}

\item{is_test}{A logical value on whether using the test-based method, default is True}

\item{nsubsamp}{A numeric value for S, the number of subsamples, default is 100}

\item{ci_level}{A numerical value corresponding to the confidence level, default is 0.95}
}
\value{
a ggplot plot object that plots the CDDR Diagnostic
}
\description{
CDDR Diagnostic: the plotted comparison of the probability of each causal discovery outcome as a function of sample size less than dataset size N.
}
\examples{

cddr <- read.csv("~/Documents/cddr_paper/package/test/cddr_test_example.csv")[,-1]
outcome_columns <- c("yonx_and_xony_reject", "yonx_and_xony_noreject",
"yonx_reject_xony_noreject", "xony_reject_yonx_noreject")
dataset_names <- c("p11", "p00", "p01", "p10")
plots <- get_diagnose_plot(cddr, outcome_columns, dataset_names)

cddr_lingam <- read.csv("~/Documents/cddr_paper/package/test/cddr_lingam_example.csv")[,-1]
outcome_columns <- c("order_right", "order_left")
dataset_names <- c("d1", "d2")
plots <- get_diagnose_plot(cddr_lingam, outcome_columns, dataset_names, is_test = FALSE)

cddr_random <- read.csv("~/Documents/cddr_paper/package/test/cddr_random.csv")[,-1]
outcome_columns <- c("order_right", "order_left")
dataset_names <- c("d1", "d2")
plots <- get_diagnose_plot(cddr_random, outcome_columns, dataset_names, is_test = FALSE)
}
