% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_cal}
\alias{pop_cal}
\title{Population and Food Consumption Data}
\format{
\subsection{\code{pop_cal}}{

A data frame with 347 rows and 2 columns:
\describe{
\item{x}{average annual rate of change in population}
\item{y}{average annual rate of change in total dietary consumption}
...
}
}
}
\source{
\url{https://webdav.tuebingen.mpg.de/cause-effect/pair0076.txt}
}
\usage{
pop_cal
}
\description{
Data collected from the Food and Agriculture Organization of the United
Nations, obtained by Mooji et al. ...
}
\keyword{datasets}
