% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_procedure.R
\name{run_procedure}
\alias{run_procedure}
\title{Get estimated causal discovery outcome rates as a function of subsample size}
\usage{
run_procedure(
  data,
  dir,
  sampsizes,
  fit_resid_yonx,
  predict_yonx,
  fit_resid_xony,
  predict_xony,
  names_col_cddr,
  name_files,
  name_out,
  run_parallel = T,
  run_lingam = F,
  clear_files = T,
  nsubsamp = 100,
  early_stop = T,
  early_stop_thresh = 2
)
}
\arguments{
\item{data}{A dataframe of size N with two columns: x and y}

\item{dir}{A string of the directory where want to save the csv file of estimated causal discovery outcome rates for each subsample size}

\item{sampsizes}{A vector of subsample sizes where the largest subsample size is less than equal to N}

\item{fit_resid_yonx}{A function that takes in a dataframe with variable x, fits a model of choice on the data (e.g. linear regression) where y is the outcome and x is the predictor, and function outputs a list with two components: model, which is the model object, and residuals, the residuals after fitting the model on the data}

\item{predict_yonx}{A function that takes in a model object (where y is the outcome and x is the predictor) and newdata with two variables x and y, gets predictions based on the model for the newdata and returns these predictions}

\item{fit_resid_xony}{A function that takes in a dataframe with variable y, fits a model of choice on the data (e.g. linear regression) where x is the outcome and y is the predictor, and function outputs a list with two components: model, which is the model object, and residuals, the residuals after fitting the model on the data}

\item{predict_xony}{A function that takes in a model object (where x is the outcome and y is the predictor) and newdata with two variables x and y, gets predictions based on the model for the newdata and returns these predictions}

\item{names_col_cddr}{A vector of names for each of the columns in the estimated causal discovery outcome rates dataframe}

\item{name_files}{A vector of names for the saved files for each subsample size (just the root name, such as cddr_samp)}

\item{name_out}{A vector of names for the output files corresponding to the estimated causal discovery outcome rates dataframe as well as the pvalues}

\item{run_parallel}{A logical value on whether to run items using parallel processing, default is True}

\item{run_lingam}{A logical value on whether to run LiNGAM method, default is False to run the test-based method}

\item{clear_files}{A logical value on whether to clear the files generated in the bootstrap procedure for each subsample size, default is True}

\item{nsubsamp}{A numeric value for S, the number of subsamples}

\item{early_stop}{A logical value on whether to stop the method early for the test-based method if the rate of rejecting both directions converges to 1, default is True}

\item{early_stop_thresh}{A numerical value on what the threshold is to stop early (e.g after 3 subsample sizes where the rate of rejecting in both directions is 1), default is 2}
}
\value{
A dataframe of the estimated causal discovery outcome rates for each subsample size considered. Also locally saves these probabilities as well as the p-values into the corresponding name_out files
}
\description{
Get estimated causal discovery outcome rates as a function of subsample size
}
\examples{
sampsizes <- c(seq(20, 40, 10))
fit_resid_yonx <- function(data) {
  model <- lm(formula("y~x"), data = data)
  resids <- residuals(model)
  model_obs <- list(model = model, residuals = resids)
  return(model_obs)
  }
predict_yonx <- function(model, data) {
  preds <- predict(model,newdata = data)
  return(preds)
  }
fit_resid_xony <- function(data) {
  model <- lm(formula("x~y"), data = data)
  resids <- residuals(model)
  model_obs <- list(model = model, residuals = resids)
  return(model_obs)
  }
predict_xony <- function(model, data) {
  preds <- predict(model,newdata = data)
  return(preds)
  }
names_col_cddr <- c("samplesizes", "yonx_and_xony_reject",
"yonx_and_xony_noreject", "yonx_reject_xony_noreject",
"xony_reject_yonx_noreject")
main_dir <- "~/Documents/cddr_paper/package/test"
name_files <- c("cddr_samp", "p_valsxy_samp", "p_valsyx_samp")
name_out <- c("cddr_test", "p_valsxy", "p_valsyx")
cddr <- run_procedure(pop_cal, main_dir, sampsizes,
fit_resid_yonx, predict_yonx, fit_resid_xony, predict_xony,
names_col_cddr, name_files, name_out)
}
